% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/input.R
\name{compute_word_duration}
\alias{compute_word_duration}
\title{How many timeslices are needed to encode a number of phonemes?}
\usage{
compute_word_duration(num_phones)
}
\description{
How many timeslices are needed to encode a number of phonemes?
}

