% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/features-to-phonemes.R
\name{FeatureDetector}
\alias{FeatureDetector}
\title{Create a pool of mutually inhibitory nodes for all values in a single
feature.}
\usage{
FeatureDetector(type, time)
}
\description{
TODO the feature value range (0 to 8) is hard-coded. Should be expressed as a
parameter.
}

