% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/input.R
\name{FeatureMatrix}
\alias{FeatureMatrix}
\title{Create an matrix to hold the feature values of an input word}
\usage{
FeatureMatrix(n_timeslices, feature_range = 0:8)
}
\description{
Create an matrix to hold the feature values of an input word
}

