% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/input.R
\name{fill_feature_matrix}
\alias{fill_feature_matrix}
\title{TODO set skirt-length as a parameter.}
\usage{
fill_feature_matrix(phoneme, phoneme_number, feat_matrix, phoneme_set)
}
\description{
TODO set skirt-length as a parameter.
}

