% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/network.R
\name{initialize_network}
\alias{initialize_network}
\title{Create a TRACE network}
\usage{
initialize_network(feature_input, lexicon)
}
\arguments{
\item{feature_input}{a matrix of feature values over time}

\item{lexicon}{a lexicon data-frame}
}
\value{
a list of interconnected feature, phoneme, and word nodes
}
\description{
A Network object is just a list of Nodes. The print function shows the number
of nodes of each type, the timespan of the network and the number of updates
applied.
}

