% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node.R
\docType{data}
\name{Node}
\alias{Node}
\title{The Node class}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x000000001369f528> 
 - attr(*, "name")= chr "Node_generator"
}}
\usage{
Node
}
\description{
The Node class
}
\keyword{datasets}

