% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/features-to-phonemes.R
\name{PhonemePool}
\alias{PhonemePool}
\title{Create two sets of overlapping phonemes}
\usage{
PhonemePool(n_timeslices)
}
\arguments{
\item{n_timeslices}{number of timeslices to span. Because phonemes are 6
timeslices wide, the number of timeslices must be divisible by 6}
}
\value{
a list of Phoneme
}
\description{
Create two sets of overlapping phonemes
}

