% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/retrace-package.r
\docType{data}
\name{phonemes86}
\alias{phonemes86}
\title{1986 Phoneme definitions}
\format{A \code{data.frame} with 98 observations.}
\usage{
phonemes86
}
\description{
1986 Phoneme definitions
}
\examples{
phonemes86
}

