% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/input.R
\name{plot_feature_input}
\alias{plot_feature_input}
\title{Plot feature values}
\usage{
plot_feature_input(feature_matrix)
}
\arguments{
\item{feature_matrix}{an input feature matrix}
}
\value{
a ggplot object showing how feature values vary and overlap over the
  course of the time input
}
\description{
Plot feature values
}

