% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/words.R
\name{spread_many_units}
\alias{spread_many_units}
\title{Spread several units (smart).}
\usage{
spread_many_units(units, n_timeslices)
}
\description{
This function is smart because it computes the spread of each unit-length
only once. Instead of doing 14 spreads for the phoneme alphabet (all 1 unit
long) for example, this function does a single spread and copies it 14 times.
}

