% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/features-to-phonemes.R
\name{spread_phonemes}
\alias{spread_phonemes}
\title{Create plan of how to spread phonemes over the two layers}
\usage{
spread_phonemes(phoneme_alphabet, n_timeslices)
}
\description{
Create plan of how to spread phonemes over the two layers
}

