% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{str_censor}
\alias{str_censor}
\title{Convert all characters in a string to hyphens}
\usage{
str_censor(xs)
}
\arguments{
\item{xs}{a character vector}
}
\value{
a vector with the same number of strings (and characters) as the
  input, but with all characters converted to hyphens
}
\description{
Convert all characters in a string to hyphens
}
\examples{
str_censor(c("", "a", "ab"))
# c("", "-", "--")
}

