% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{str_inventory}
\alias{str_inventory}
\title{Make an inventory of characters in a string}
\usage{
str_inventory(word)
}
\arguments{
\item{word}{a string}
}
\value{
a sorted vector of unique characters in the string
}
\description{
Make an inventory of characters in a string
}
\examples{
str_inventory("letters")
# c("e", "l", "r", "s", "t")
}

