% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{str_tokenize}
\alias{str_tokenize}
\title{Convert a string into a vector of characters}
\usage{
str_tokenize(word)
}
\arguments{
\item{word}{a string}
}
\value{
a vector of characters in the string
}
\description{
Convert a string into a vector of characters
}
\examples{
str_tokenize("letters")
# c("l", "e", "t", "t", "e", "r", "s")
}

