% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{str_wrap_silence}
\alias{str_wrap_silence}
\title{Add silence to start and end of word}
\usage{
str_wrap_silence(xs)
}
\arguments{
\item{xs}{a character vector}
}
\value{
the inputted strings but with silence characters ("-") added to
  beginning and end of word.
}
\description{
Add silence to start and end of word
}
\examples{
str_wrap_silence(c("", "a", "ab"))
# c("--", "-a-", "-ab-")
}

