% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edge.R
\name{visit}
\alias{visit}
\title{Visit a Node over an Edge}
\usage{
visit(x, f = multiply_by)
}
\arguments{
\item{x}{an Edge or a list of Edges.}

\item{f}{a binary function used to combine the visited node's activation and
the edge's weight. Default is multiplication, so f(activation, weight) =
activation * weight.}
}
\value{
a vector of f(activation, weight) values
}
\description{
Visit a Node over an Edge
}

